Split Database.kt into two files.
